+++ backend/src/main/java/ch/cevi/db/adapter/hitobito/HitobitoProvider.java
@@ -1,22 +1,15 @@
 package ch.cevi.db.adapter.hitobito;
 
+import ch.cevi.db.adapter.EventFilter;
 import ch.cevi.db.adapter.domain.CeviEvent;
-import ch.cevi.db.adapter.domain.CeviEventType;
 import ch.cevi.db.adapter.domain.Kursart;
 import ch.cevi.db.adapter.domain.Organisation;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
 public interface HitobitoProvider {
-    List<CeviEvent> getEvents(Optional<String> groupFilter,
-                              Optional<LocalDate> earliestStartAt,
-                              Optional<LocalDate> lastStartAt,
-                              Optional<String> nameContains,
-                              Optional<CeviEventType> eventType,
-                              Optional<String> kursartFilter);
+    List<CeviEvent> getEvents(EventFilter filter);
     List<Organisation> getOrganisations();
     List<Kursart> getKursarten();
     long getAnzahlAnlaesse();