+++ backend/src/test/java/ch/cevi/db/adapter/EventControllerIT.java
@@ -16,7 +16,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SpringBootTest
@@ -32,14 +32,13 @@ class EventControllerIT {
     @Test
     void should_retrieve_events() throws Exception {
         String content = mockMvc.perform(
-                get("/events")
+                post("/events")
+                        .content(objectMapper.writeValueAsString(EventFilter.emptyFilter()))
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andReturn().getResponse().getContentAsString(StandardCharsets.UTF_8);
         List<CeviEvent> events = objectMapper.readValue(content, new TypeReference<>() {});
         assertThat(events).hasSizeGreaterThan(20);
         assertThat(events).allMatch(e -> e.startsAt().isAfter(LocalDate.now().atStartOfDay()));
-        assertThat(events.stream().filter(e -> e.id().equals("3821")).findFirst().orElseThrow().name()).isEqualTo("YMCA Europe / General Assembly");
-        assertThat(events.stream().filter(e -> e.id().equals("3860")).findFirst().orElseThrow().name()).isEqualTo("Gruppenleiterkurs GLK 2024");
     }
 }